Make links show exact amount of goals
